Add a ton of debugging statements together with "debug" configuration setting
It turned out that one simply needs debugging statements when things don't work as expected.
Before everyone has to add these himself, we add them once and for all, accepting the cost of
blowing up the source by a fair amount. Simply enable the "debug" option in the plugin's
configuration.