Bump date in plugin.info.txt for consistent update process
Since "Ponder Stibbons", DokuWiki can notify about updates to plugins. It
does so by comparing the date of the (local) plugin.info.txt with the date
on the plugin's page. We updated the latter but not the former, so we'd
at least have to make these two the same.

As, however, the plugin also shows the latest Git commits, including this
one from today, it makes sense to just bump the date to today.